+++ docs/fr_FR/changelog.md
@@ -1,3 +1,8 @@
+# 09/05/2020
+
+- Ajout du support de l'extension 0-10v
+- optimisation
+
 # 30/12/2019
 
 - Correction bug thermostat